requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Session==0.5.0
python-dotenv==1.0.1
requests==2.32.3
google-auth-oauthlib==1.2.1
pytesseract==0.3.10
Pillow==10.4.0
opencv-python==4.9.0.80
APScheduler==3.10.4
numpy<2.0
